Add npm command to docker compose file
Previously, the docker compose file did not run npm install, which
would cause the application to fail for new users. This has been
fixed so that 'npm install' is run on docker-compose up.

In addition, the package-lock file has been updated from file
version 1 to file version 2. This was a result of using a newer
version of npm from the docker image, rather than the one installed
on my latest Fedora 33 dev laptop. I wasn't able to find much
information on what this actually is, so I'm assuming newer =
better in this instance.