Add new command to add a note
Previously, the application was doing a mish-mash of stuff that
I had added from going through the lesson and tweaking as I saw
fit. Now that we're moving into the actual notes portion of the
application, it's time to change some things.

There is now a command that will add a note to the given file,
which can be seen from --help. Note that the code to do this
is different than what's shown in the course, I'm guessing
that yargs had a major update since then and what's shown
in the video doesn't work anymore.