Add all commands to help menu
Previously, the new commands were not showing in the help menu.
After some digging in the yargs docs, I found I was using
the wrong syntax for multiple commands, and needed to chain
them together. It is now working.

Also updated the changelog again. Hopefully didn't miss anything
this time.